+++ src/app/cartography/events/nodes.ts
@@ -1,0 +1,14 @@
+import { Node } from "../models/node";
+
+class NodeEvent {
+    constructor(
+        public event: any,
+        public node: Node
+    ) {}
+}
+
+export class NodeDragging extends NodeEvent {}
+export class NodeDragged extends NodeEvent {}
+
+export class NodeClicked extends NodeEvent {}
+export class NodeContextMenu extends NodeEvent {}
